Fix the ssh command at the bottom of the setup script.
It turns out that using gcloud cloud-shell ssh rewrites the authorized
keys which are stored in /etc/ssh/keys/authorized_keys and undoes the
setup the script had done to put our usual keys into that location.

Avoid using gcloud cloud-shell for that reason, but the implication is
that --authorize_session isn't passed to it so gcloud won't then work
inside the cloudshell VM with your users' credentials. At the moment we
don't need that so it's fine, but perhaps that will cause an issue later.