Remove usage of Google cloud's identity file.
In the original ssh config output, an IdentityFile line was being
generated for the owner of the cloud shell which would use your
Google Cloud ssh keys to establish identity with the cloud shell.

This won't work for collaborators using the config as the script
tells you to do, and isn't necessary for your own identity either,
so simply remove it.