A working development container.
In this Dockerfile approach, the docker image runs an ssh daemon
that you can connect to via a ProxyJump through the host that runs
the docker container.

Then inside the development container, all collaborators can run
tmux in order to be in a shared terminal.

You can exit the container for everyone by running

docker kill ot

from within the container. The container has access to the containing
computer's docker socket, chmod'ed to allow the dev user to speak to
it directly. This constitutes a security hole from the perspective of
the containing OS because docker access is normally reserved for root,
so ensure that it's only you in the container and/or that the containing
system is a VM you don't care much about.