allow operator's command to be different from the OS executable
Want to explore what things might look like if we let the operator "commands"
(which should likely be renamed to "names") be different from the actual
operating system commands used to perform the operations.

Testing this approach by dropping the `bin/dd status=none` and
`bin/dd status=none if=%{dir}%{file}` operators, and replacing them
with `bin/store` and `bin/fetch`, respectively, while still using
the OS's `dd` executable as the actual implementation.

To avoid adding more config to the `operators.yml` file, we're adding
a new argument to the operator constructor (`cmd_to_exec`) to allow
for this divergence.  In order to achieve backwards compatibility,
the value defaults to `cmd`, when not provided.