simplify test pipeline
During our last code review, we tried to remove the `tr` portion
of the pipeline, wanting to rely on awk's "printf %s".

Replacing

  let(:op_pipeline) { ['wc', "awk '{print $%<test_arg>s}'", 'tr -d "\n"'] }

with

  let(:op_pipeline) { ['wc', "awk '{printf \"%s\", $%<test_arg>s}'"] }

caused tests to seemingly break.

The problem, however, was not due to the production code, but to the test
change itself.  In particular, the "%" in "%s" was being interpreted by
the argument replacement operator (also "%") as a portion of the string
to be replaced, but for which a replacement value had not been supplied.

The "%" in "%s" needed to be escaped using "%%s", i.e.

  let(:op_pipeline) { ['wc', "awk '{printf \"%%s\", $%<test_arg>s}'"] }

To recap, in ruby:
  "%s" % {x: 1}
  => "{:x=>1}" # the %s is replaced by the stringified hash
  "%s" % 1
  => "1"
  "%s %<x>s" % 1
  => # error: `%': named<x> after unnumbered(1)
  "%%s"
  => "%s" # this is what we wanted