add flexibility to Cmds::Generic
Previously Operator's exec() methods always returned a string: the
output of the executed command.

Sometime, however, the Operator generates multiple outputs (e.g.
Split, which generates a series of files with partial content),
and we may want to serialize multiple inverse operators based on
this series of outputs.

For example, we want to support the ability for the
Split command to have a fwd_op that creates multiple bits of content,
allowing then for the production of multiple serialized inv_op's,
one for each piece of content.

$echo "hello world" | bin/split -b 5
>><<catcathello>><<catcat worl>><<catcatd

To allow this, a new `fwd_op_proc` argument is allowed to be passed
to `exec`.  This Proc receives fwd_op's output as its sole argument
and can return a modified output (including an array, as in the case
of Split's fwd_op, which returns an array of filenames that have
been produced).

Cmds::Generic's exec() then iterates over this result and produces
an inverse operator for each one.

Because of this change, since Ruby methods cannot accept more than
one implitic blocks (i.e. invocable via `yield`), and since exec()
already accepted an implicit block to transform the content to
serialize, the method signature was modified to accept both:

- fwd_op_proc: a Proc as described above
- serialization_modifier_proc: a Proc for potential modifications to
  each piece of content to be serialized

With these new capabilities of Cmds::Generic, it seemed also cleaner
to remove several of the other Cmds:: classes, in favour of the
(small amount of) code being directly in the corresponding bin files.

Tests were correspondigly updated, with the following additional non-obvious
change.  The `be_operator` matcher was augmented to accept
`.allowing_additional_content`, and the matcher now fails
if the content being tested has extra traling content.  This
feature is particularly helpful in testing the output of commands
like Split, which produce a sequence of operators.  Testing some
parts of Split can now look simply like this:

  expect(output_stream).to be_operarator('cat')
    ...
    .allowing_additional_content
  expect(output_stream).to be_operarator('cat')
    ...
    .allowing_additional_content
  expect(output_stream).to be_operarator('cat')
    ...

It's important to note that for this feature to work, we need
to use streams, not simple strings, because we need each matcher
to keep reading from where the previous one left off.  Though
I'm not 100% happy with this DSL and may change it later, it
does prove useful for the time being.