Setup script for hosting a shared shell environment.
This setup script uses the user's cloud shell to build a docker
container in which a group of people can work on the code in a
tmux session.

The docker container gets rebuilt each time (unless you've run
the command recently, in which case your cloud shell will still
have a copy of the last image you built).

For now it outputs a configuration file to the clipboard on macos
that can be sent to a collaborator to allow them to join the
session.

An issue that seems to plague the cloud shell CLI is that only
the cloud-shell ssh command itself seems to check for and spin
up the cloud shell instance. So, start cloud shell before doing
other operations by running an echo command on the remote VM.