Add .bundle/config to the project.
In the absence of a .bundle/config in the project, executing

bundle install

will attempt to install ruby gems globally which the user won't (and
shoudln't) have permission to do. Including a .bundle/config file
in the project means that ruby gems will get installed local to this
project only and with the user's permissions avoiding the problem.